Splitted priv.h into many different header files.
This allows to decouple some deps and improve code readability.
