+++ Lib/evts.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "src.h"
+
+/* Struct that holds an event + its source, private */
+typedef struct {
+    m_evt_t evt;
+    ev_src_t *src;                          // Ref to src that caused the event
+} evt_priv_t;
+
+evt_priv_t *new_evt(ev_src_t *src);
+
