+++ Lib/ps.c
@@ -1,5 +1,7 @@
-#include "mod.h"
+#include "ps.h"
 #include "fs_priv.h"
+#include "evts.h"
+#include "ctx.h"
 
 /******************************************
  * Code related to Actor-like PubSub API. *
