+++ Lib/public/module/list.h
@@ -1,10 +1,5 @@
 #pragma once
 
-#ifndef LIBMODULE_STRUCT_H
-    #define LIBMODULE_STRUCT_H
-#endif
-#include "cmn.h"
-
 /** Linked-List interface **/
 
 /* Callback for list_iterate; first parameter is userdata, second is list data */
