+++ Lib/public/module/map.h
@@ -1,9 +1,6 @@
 #pragma once
 
-#ifndef LIBMODULE_STRUCT_H
-    #define LIBMODULE_STRUCT_H
-#endif
-#include "cmn.h"
+#include <stdbool.h>
 
 /** Hashmap interface **/
