+++ Lib/utils/mem.c
@@ -1,3 +1,4 @@
+#include "public/module/mem.h"
 #include "priv.h"
 #include <stddef.h>
 #include <stdalign.h>
@@ -18,6 +19,18 @@ static inline mem_header_t *get_header(uint8_t *src) {
 }
 
 /** Private API **/
+
+char *mem_strdup(const char *s) {
+    char *new = NULL;
+    if (s) {
+        const size_t len = strlen(s) + 1;
+        new = memhook._malloc(len);
+        if (new) {
+            memcpy(new, s, len);
+        }
+    }
+    return new;
+}
 
 void mem_dtor(void *src) {
     m_mem_unref(src);
