+++ Lib/utils/utils.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <stdint.h>
+#include <stdbool.h>
+
+void fetch_ms(uint64_t *val, uint64_t *ctr);
+bool str_not_empty(const char *str);
